backend/requirements.txt
fastapi
uvicorn
python-dotenv

aiohttp
google-genai
Pillow
supabase
tavily-python


watchdog
